Create and save walking session to DB

+++ app/src/main/java/it/polimi/steptrack/roomdatabase/entities/AccelerometerSample.java
@@ -20,7 +20,7 @@ public class AccelerometerSample {
     public long AsTimestamp;
 
     @ColumnInfo(name = "session_id")
-    public int SessionID;
+    public long SessionID;
 
     @ColumnInfo(name = "acc_x")
     public float mAccX;
